+++ src/routing/surface_kinwave.jl
@@ -28,13 +28,13 @@ end
 "Initialize variables for river or overland flow models"
 function FlowVariables(n)
     variables = FlowVariables(;
-        q = zeros(FLOAT, n),
-        qlat = zeros(FLOAT, n),
-        qin = zeros(FLOAT, n),
-        q_av = zeros(FLOAT, n),
-        volume = zeros(FLOAT, n),
-        h = zeros(FLOAT, n),
-        h_av = zeros(FLOAT, n),
+        q = zeros(Float64, n),
+        qlat = zeros(Float64, n),
+        qin = zeros(Float64, n),
+        q_av = zeros(Float64, n),
+        volume = zeros(Float64, n),
+        h = zeros(Float64, n),
+        h_av = zeros(Float64, n),
     )
     return variables
 end
@@ -55,12 +55,12 @@ end
 function ManningFlowParameters(slope, mannings_n, flow_length, flow_width)
     n = length(slope)
     parameters = ManningFlowParameters(;
-        beta = FLOAT(0.6),
+        beta = Float64(0.6),
         slope,
         mannings_n,
         flow_length,
         flow_width,
-        alpha_pow = FLOAT((2.0 / 3.0) * 0.6),
+        alpha_pow = Float64((2.0 / 3.0) * 0.6),
         alpha_term = fill(MISSING_VALUE, n),
         alpha = fill(MISSING_VALUE, n),
     )
@@ -92,7 +92,7 @@ function RiverFlowParameters(dataset, config, indices, river_length, river_width
         "routing.river_flow.mannings_n";
         sel = indices,
         defaults = 0.036,
-        type = FLOAT,
+        type = Float64,
     )
     bankfull_depth = ncread(
         dataset,
@@ -101,7 +101,7 @@ function RiverFlowParameters(dataset, config, indices, river_length, river_width
         alias = "routing.river_flow.h_bankfull",
         sel = indices,
         defaults = 1.0,
-        type = FLOAT,
+        type = Float64,
     )
     if haskey(config.input.routing.river_flow, "h_bankfull")
         @warn string(
@@ -115,7 +115,7 @@ function RiverFlowParameters(dataset, config, indices, river_length, river_width
         "routing.river_flow.slope";
         optional = false,
         sel = indices,
-        type = FLOAT,
+        type = Float64,
     )
     clamp!(slope, 0.00001, Inf)
 
@@ -138,10 +138,10 @@ end
 "Initialize river flow model boundary conditions"
 function RiverFlowBC(n, reservoir, lake)
     bc = RiverFlowBC(;
-        inwater = zeros(FLOAT, n),
-        inflow = zeros(FLOAT, n),
-        inflow_waterbody = zeros(FLOAT, n),
-        abstraction = zeros(FLOAT, n),
+        inwater = zeros(Float64, n),
+        inflow = zeros(Float64, n),
+        inflow_waterbody = zeros(Float64, n),
+        abstraction = zeros(Float64, n),
         reservoir = reservoir,
         lake = lake,
     )
@@ -173,7 +173,7 @@ function KinWaveRiverFlow(
         init_kinematic_wave_timestepping(config, n; domain = "river", dt_fixed = 900.0)
 
     do_water_demand = haskey(config.model, "water_demand")
-    allocation = do_water_demand ? AllocationRiver(n) : NoAllocationRiver{FLOAT}()
+    allocation = do_water_demand ? AllocationRiver(n) : NoAllocationRiver{Float64}()
 
     variables = FlowVariables(n)
     parameters = RiverFlowParameters(dataset, config, indices, river_length, river_width)
@@ -228,16 +228,16 @@ function KinWaveOverlandFlow(dataset, config, indices; slope, flow_length, flow_
         "routing.overland_flow.mannings_n";
         sel = indices,
         defaults = 0.072,
-        type = FLOAT,
+        type = Float64,
     )
 
     n = length(indices)
     timestepping =
         init_kinematic_wave_timestepping(config, n; domain = "land", dt_fixed = 3600.0)
 
-    variables = LandFlowVariables(; flow = FlowVariables(n), to_river = zeros(FLOAT, n))
+    variables = LandFlowVariables(; flow = FlowVariables(n), to_river = zeros(Float64, n))
     parameters = ManningFlowParameters(slope, mannings_n, flow_length, flow_width)
-    boundary_conditions = LandFlowBC(; inwater = zeros(FLOAT, n))
+    boundary_conditions = LandFlowBC(; inwater = zeros(Float64, n))
     sf_land =
         KinWaveOverlandFlow(; timestepping, boundary_conditions, variables, parameters)
